add ResourceLoader
