+++ projects/kiwano/kiwano.vcxproj
@@ -102,6 +102,7 @@
     <ClInclude Include="..\..\src\kiwano\utils\Json.h" />
     <ClInclude Include="..\..\src\kiwano\utils\Logger.h" />
     <ClInclude Include="..\..\src\kiwano\utils\ResourceCache.h" />
+    <ClInclude Include="..\..\src\kiwano\utils\ResourceLoader.h" />
     <ClInclude Include="..\..\src\kiwano\utils\Task.h" />
     <ClInclude Include="..\..\src\kiwano\utils\TaskScheduler.h" />
     <ClInclude Include="..\..\src\kiwano\utils\Ticker.h" />
@@ -179,6 +180,7 @@
     <ClCompile Include="..\..\src\kiwano\utils\EventTicker.cpp" />
     <ClCompile Include="..\..\src\kiwano\utils\Logger.cpp" />
     <ClCompile Include="..\..\src\kiwano\utils\ResourceCache.cpp" />
+    <ClCompile Include="..\..\src\kiwano\utils\ResourceLoader.cpp" />
     <ClCompile Include="..\..\src\kiwano\utils\Task.cpp" />
     <ClCompile Include="..\..\src\kiwano\utils\TaskScheduler.cpp" />
     <ClCompile Include="..\..\src\kiwano\utils\Ticker.cpp" />
