+++ projects/kiwano/kiwano.vcxproj.filters
@@ -354,6 +354,9 @@
     <ClInclude Include="..\..\src\kiwano\base\RefPtr.h">
       <Filter>base</Filter>
     </ClInclude>
+    <ClInclude Include="..\..\src\kiwano\utils\ResourceLoader.h">
+      <Filter>utils</Filter>
+    </ClInclude>
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="..\..\src\kiwano\2d\Canvas.cpp">
@@ -577,6 +580,9 @@
     </ClCompile>
     <ClCompile Include="..\..\src\kiwano\base\RefObject.cpp">
       <Filter>base</Filter>
+    </ClCompile>
+    <ClCompile Include="..\..\src\kiwano\utils\ResourceLoader.cpp">
+      <Filter>utils</Filter>
     </ClCompile>
   </ItemGroup>
   <ItemGroup>
