+++ src/kiwano/utils/ResourceLoader.h
@@ -19,59 +19,37 @@
 // THE SOFTWARE.
 
 #pragma once
-#include <kiwano/core/Resource.h>
-#include <kiwano/render/Frame.h>
-#include <kiwano/render/FrameSequence.h>
-#include <kiwano/render/Font.h>
-#include <kiwano/render/GifImage.h>
+#include <kiwano/utils/ResourceCache.h>
+#include <kiwano/utils/Json.h>
+#include <kiwano/utils/Xml.h>
 
 namespace kiwano
 {
 
-KGE_DECLARE_SMART_PTR(ResourceCache);
-
 /// \~chinese
-/// @brief 资源缓存
-class KGE_API ResourceCache final : public ObjectBase
+/// @brief 资源加载器
+class KGE_API ResourceLoader final : Noncopyable
 {
 public:
-    ResourceCache();
+    /// \~chinese
+    /// @brief 从 JSON 文件加载资源信息
+    /// @param file_path JSON文件路径
+    static bool LoadFromJsonFile(ResourceCachePtr cache, const String& file_path);
 
     /// \~chinese
-    /// @brief 获取资源
-    /// @param id 对象ID
-    ObjectBasePtr Get(const String& id) const;
+    /// @brief 从 JSON 加载资源信息
+    /// @param json_data JSON对象
+    static bool LoadFromJson(ResourceCachePtr cache, const Json& json_data);
 
     /// \~chinese
-    /// @brief 获取资源
-    /// @tparam _Ty 对象类型
-    /// @param id 对象ID
-    /// @return 指定对象类型的指针
-    template <typename _Ty>
-    RefPtr<_Ty> Get(const String& id) const
-    {
-        return dynamic_cast<_Ty*>(Get(id).Get());
-    }
+    /// @brief 从 XML 文件加载资源信息
+    /// @param file_path XML文件路径
+    static bool LoadFromXmlFile(ResourceCachePtr cache, const String& file_path);
 
     /// \~chinese
-    /// @brief 将对象放入缓存
-    /// @param id 对象ID
-    /// @param obj 对象
-    void AddObject(const String& id, ObjectBasePtr obj);
-
-    /// \~chinese
-    /// @brief 删除指定资源
-    /// @param id 对象ID
-    void Remove(const String& id);
-
-    /// \~chinese
-    /// @brief 清空所有资源
-    void Clear();
-
-    virtual ~ResourceCache();
-
-private:
-    UnorderedMap<String, ObjectBasePtr> object_cache_;
+    /// @brief 从 XML 文档对象加载资源信息
+    /// @param doc XML文档对象
+    static bool LoadFromXml(ResourceCachePtr cache, const XmlDocument& doc);
 };
 
 }  // namespace kiwano
